refactor(web): move and update AboutSection component
Move AboutSection from components to sections and update its
implementation for improved structure and usability.
Export AboutSection via sections/index.ts for better code organization
and reusability.
