+++ apps/web/src/app/sections/index.ts
@@ -1,1 +1,2 @@
 export * from './Hero';
+export * from './About';
